# design_docs/openapi.v0.yaml
openapi: 3.0.3
info:
  title: Curious Now API (v0)
  version: "0.1"
servers:
  - url: /v1
paths:
...
  /feed:
    get:
      summary: Stage 2+ cluster feed (latest/trending/for_you)
      parameters:
        - in: query
          name: tab
          schema:
            type: string
            enum: [latest, trending, for_you]
            default: latest
        - in: query
          name: topic_id
          schema: { type: string, format: uuid }
        - in: query
          name: content_type
          schema: { $ref: "#/components/schemas/ContentType" }
        - in: query
          name: page
          schema: { type: integer, minimum: 1, default: 1 }
        - in: query
          name: page_size
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
      responses:
        "200":
          description: Paginated list of StoryClusters (cards)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClustersFeedResponse"
        "401":
          description: Unauthenticated (only returned when tab=for_you)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /clusters/{id}:
    get:
      summary: Get a StoryCluster and its evidence list
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Cluster detail
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ClusterDetail" }
        "301":
          description: Cluster has been merged; client should redirect to canonical cluster id
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RedirectResponse" }
  /clusters/{id}/updates:
    get:
      summary: Stage 4 update log for a cluster
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
        - in: query
          name: page
          schema: { type: integer, minimum: 1, default: 1 }
        - in: query
          name: page_size
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
      responses:
        "200":
          description: Cluster updates
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ClusterUpdatesResponse" }
        "301":
          description: Cluster has been merged; client should redirect to canonical cluster id
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RedirectResponse" }
...
  /topics/{id}:
    get:
      summary: Topic page (clusters in topic)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Topic detail
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TopicDetail" }
        "301":
          description: Topic has been merged; client should redirect to canonical topic id
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TopicRedirectResponse" }
  /topics/{id}/lineage:
    get:
      summary: Stage 4 lineage nodes/edges for a topic
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Lineage graph for topic
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TopicLineageResponse" }
        "301":
          description: Topic has been merged; client should redirect to canonical topic id
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TopicRedirectResponse" }
  /search:
    get:
      summary: Stage 2+ search (cluster-first)
      parameters:
        - in: query
          name: q
          required: true
          schema: { type: string, minLength: 1 }
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SearchResponse" }
  /glossary:
    get:
      summary: Stage 3 glossary lookup (term or alias)
      parameters:
        - in: query
          name: term
          required: true
          schema: { type: string, minLength: 1 }
      responses:
        "200":
          description: Best matching glossary entry
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GlossaryLookupResponse" }
...
components:
  securitySchemes:
    AdminTokenAuth:
      type: apiKey
      in: header
      name: X-Admin-Token
    SessionAuth:
      type: apiKey
      in: cookie
      name: cn_session
  schemas:
    ContentType:
      type: string
      enum: [news, press_release, preprint, peer_reviewed, report]
    EntityType:
      type: string
      enum: [person, institution, model, dataset, method, venue]
...
    TopicChip:
      type: object
      required: [topic_id, name, score]
      properties:
        topic_id: { type: string, format: uuid }
        name: { type: string }
        score: { type: number }
    ClusterCard:
      type: object
      required: [cluster_id, canonical_title, updated_at, distinct_source_count]
      properties:
        cluster_id: { type: string, format: uuid }
        canonical_title: { type: string }
        updated_at: { type: string, format: date-time }
        distinct_source_count: { type: integer }
        top_topics:
          type: array
          items: { $ref: "#/components/schemas/TopicChip" }
        content_type_badges:
          type: array
          items: { $ref: "#/components/schemas/ContentType" }
        method_badges:
          type: array
          items: { type: string }
        takeaway: { type: string, nullable: true }
        confidence_band:
          type: string
          nullable: true
          enum: [early, growing, established]
        anti_hype_flags:
          type: array
          items: { type: string }
        is_saved: { type: boolean, nullable: true }
        is_watched: { type: boolean, nullable: true }
        why_in_feed: { $ref: "#/components/schemas/WhyInFeed" }
    ClustersFeedResponse:
      type: object
      required: [tab, page, results]
      properties:
        tab: { type: string }
        page: { type: integer }
        results:
          type: array
          items: { $ref: "#/components/schemas/ClusterCard" }
    EvidenceItem:
      type: object
      required: [item_id, title, url, source, published_at]
      properties:
        item_id: { type: string, format: uuid }
        title: { type: string }
        url: { type: string }
        published_at: { type: string, format: date-time, nullable: true }
        source:
          type: object
          required: [source_id, name]
          properties:
            source_id: { type: string, format: uuid }
            name: { type: string }
        content_type: { $ref: "#/components/schemas/ContentType" }
    ClusterDetail:
      type: object
      required: [cluster_id, canonical_title, created_at, updated_at, distinct_source_count, evidence]
      properties:
        cluster_id: { type: string, format: uuid }
        canonical_title: { type: string }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        distinct_source_count: { type: integer }
        topics:
          type: array
          items: { $ref: "#/components/schemas/TopicChip" }
        content_type_breakdown:
          type: object
          additionalProperties: { type: integer }
        evidence:
          type: object
          additionalProperties:
            type: array
            items: { $ref: "#/components/schemas/EvidenceItem" }
        takeaway: { type: string, nullable: true }
        summary_intuition: { type: string, nullable: true }
        summary_deep_dive: { type: string, nullable: true }
        assumptions:
          type: array
          items: { type: string }
        limitations:
          type: array
          items: { type: string }
        what_could_change_this:
          type: array
          items: { type: string }
        confidence_band:
          type: string
          nullable: true
          enum: [early, growing, established]
        method_badges:
          type: array
          items: { type: string }
        anti_hype_flags:
          type: array
          items: { type: string }
        takeaway_supporting_item_ids:
          type: array
          items: { type: string, format: uuid }
        summary_intuition_supporting_item_ids:
          type: array
          items: { type: string, format: uuid }
        summary_deep_dive_supporting_item_ids:
          type: array
          items: { type: string, format: uuid }
        glossary_entries:
          type: array
          items: { $ref: "#/components/schemas/GlossaryEntry" }
        is_saved: { type: boolean, nullable: true }
        is_watched: { type: boolean, nullable: true }
    Topic:
      type: object
      required: [topic_id, name]
      properties:
        topic_id: { type: string, format: uuid }
        name: { type: string }
        description_short: { type: string, nullable: true }
        is_followed: { type: boolean, nullable: true }
...
    TopicDetail:
      allOf:
        - $ref: "#/components/schemas/Topic"
        - type: object
          required: [latest_clusters]
          properties:
            latest_clusters:
              type: array
              items: { $ref: "#/components/schemas/ClusterCard" }
            trending_clusters:
              type: array
              items: { $ref: "#/components/schemas/ClusterCard" }
    ClusterUpdateEntry:
      type: object
      required: [created_at, change_type, summary, supporting_item_ids]
      properties:
        created_at: { type: string, format: date-time }
        change_type:
          type: string
          enum: [new_evidence, refinement, contradiction, merge, split, quarantine, unquarantine, correction]
        summary: { type: string }
        diff:
          type: object
          nullable: true
          properties:
            previously:
              type: array
              items: { type: string }
            now:
              type: array
              items: { type: string }
            because:
              type: array
              items: { type: string }
        supporting_item_ids:
          type: array
          items: { type: string, format: uuid }
    ClusterUpdatesResponse:
      type: object
      required: [cluster_id, updates]
      properties:
        cluster_id: { type: string, format: uuid }
        updates:
          type: array
          items: { $ref: "#/components/schemas/ClusterUpdateEntry" }
    LineageNode:
      type: object
      required: [node_id, node_type, title]
      properties:
        node_id: { type: string, format: uuid }
        node_type:
          type: string
          enum: [paper, model, dataset, method]
        title: { type: string }
        external_url: { type: string, nullable: true }
        published_at: { type: string, format: date-time, nullable: true }
    LineageEdge:
      type: object
      required: [from, to, relation_type, evidence_item_ids]
      properties:
        from: { type: string, format: uuid }
        to: { type: string, format: uuid }
        relation_type:
          type: string
          enum: [extends, improves, compresses, replaces_in_some_settings, contradicts, orthogonal]
        evidence_item_ids:
          type: array
          items: { type: string, format: uuid }
        notes_short: { type: string, nullable: true }
    TopicLineageResponse:
      type: object
      required: [topic_id, nodes, edges]
      properties:
        topic_id: { type: string, format: uuid }
        nodes:
          type: array
          items: { $ref: "#/components/schemas/LineageNode" }
        edges:
          type: array
          items: { $ref: "#/components/schemas/LineageEdge" }
    SearchResponse:
      type: object
      required: [query, clusters]
      properties:
        query: { type: string }
        clusters:
          type: array
          items: { $ref: "#/components/schemas/ClusterCard" }
        topics:
          type: array
          items: { $ref: "#/components/schemas/Topic" }
        entities:
          type: array
          items: { $ref: "#/components/schemas/Entity" }
    RedirectResponse:
      type: object
      required: [redirect_to_cluster_id]
      properties:
        redirect_to_cluster_id: { type: string, format: uuid }
    TopicRedirectResponse:
      type: object
      required: [redirect_to_topic_id]
      properties:
        redirect_to_topic_id: { type: string, format: uuid }
...
    GlossaryEntry:
      type: object
      required: [glossary_entry_id, term, definition_short]
      properties:
        glossary_entry_id: { type: string, format: uuid }
        term: { type: string }
        definition_short: { type: string }
        definition_long: { type: string, nullable: true }
    GlossaryLookupResponse:
      type: object
      required: [entry]
      properties:
        entry: { $ref: "#/components/schemas/GlossaryEntry" }
...
    WhyInFeed:
      type: object
      properties:
        reasons:
          type: array
          items: { type: string }
        matched_topic_ids:
          type: array
          items: { type: string, format: uuid }
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
...
    Entity:
      type: object
      required: [entity_id, entity_type, name]
      properties:
        entity_id: { type: string, format: uuid }
        entity_type: { $ref: "#/components/schemas/EntityType" }
        name: { type: string }
        description_short: { type: string, nullable: true }
        external_url: { type: string, nullable: true }
...